dts: xiphos: add dts files for the Q7-revB
We need to find a better way of changing the RAM size for space
configurations.
Having two different dts files in logic-q7 is cumbersome and seems way
overkill for a oneline change!

I've tested different stategies but ideally, the board file would use a
define like in the board.c file to switch to a space configuration.

Signed-off-by: Liam Beguin <lvb@xiphos.com>