cmd: introduce xscinfo command
Add a new sub-system command to interact with a xscinfo partition.
This partition is defined as 2 sectors of QSPI flash. The first sector
holds read-only information about a given unit such as it's serial
number, MAC address, ... The second sector is read-write accessible and
can be used to store any variable to persistent memory.

More information on this sub-system can be found in the Q8 design document.

Signed-off-by: Liam Beguin <lvb@xiphos.com>