cmd: xscinfo: add option to enable writes to RO section
While setting up a new board or during development, it can be useful to
have write access to the xscinfo read-only partition.
This adds the XSCINFO_ENABLE_RO_WRITE Kconfig option to do just that.

When enabled, adding a `~` at the beginning of the variable name will
indicate the user wants to write to the read-only section.

Enable this feature by default on all Q7 units, can be disabled later
when the development slows down

Signed-off-by: Liam Beguin <lvb@xiphos.com>