board: xiphos: q7: add support for loading files from ubifs
On the new revision of the Q7 image, we intend to load files from a
ubifs partition. This allows U-Boot to detect and mount the rootfs
partition before loading files from it.

for this to work, the following need to be defined:
 - MTDIDS_DEFAULTS, MTDPARTS_DEFAULTS
 - CONFIG_MTD_PARTITIONS
 - CONFIG_CMD_UBIFS
 - CONFIG_SPI_FLASH_MTD

While at it, this also disables DFU support.

Signed-off-by: Liam Beguin <lvb@xiphos.com>