board: xiphos: q7: update boot sequence
When booting, the zynq should query the PA3 to load the appropriate copy
of it's system image (kernel, devicetree, rootfs).

This enables U-Boot to read the CopyID from the PA3 and sets the
`copyid` environment variable accordingly (nom0, nom1, gold0, gold1).
This variable is then used to load the devicetree and kernel from
flash.

Signed-off-by: Liam Beguin <lvb@xiphos.com>