board: xiphos: Q8: add logic to get active copy
The ProAsic3 selects the copy to be booted by the zynq.
On the Q7, this used to be done by reading the PA3 status register.
This cannot be done on the Q8 since the PL logic isn't loaded when
U-Boot starts. Implement the new logic which consists of reading MIO38
twice, once before the first config_done to get the active segment and once
after to get the active chip.
Also, move custom init code to `board_late_init` so that we can update
the environment variables.

Signed-off-by: Liam Beguin <lvb@xiphos.com>